Move BiomeSeedProvider implementation to the correct MixinClientWorld, to avoid crashing when disabling the fast_biome_colors feature.
